+++ pybamm/models/submodels/interface/kinetics/tafel.py
@@ -30,11 +30,22 @@ class ForwardTafel(BaseKinetics):
     def __init__(self, param, domain, reaction, options):
         super().__init__(param, domain, reaction, options)
 
+    @property
+    def alpha(self):
+        if self.domain == "Negative":
+            return self.param.alpha_bv_n
+        elif self.domain == "Positive":
+            return self.param.alpha_bv_p
+
     def _get_kinetics(self, j0, ne, eta_r, T, u):
-        return u * j0 * pybamm.exp((ne / (2 * (1 + self.param.Theta * T))) * eta_r)
+        alpha = self.alpha
+        return (
+            u * j0 * pybamm.exp((ne * alpha / (2 * (1 + self.param.Theta * T))) * eta_r)
+        )
 
     def _get_dj_dc(self, variables):
         """See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_dc`"""
+        alpha = self.alpha
         (
             c_e,
             delta_phi,
@@ -46,17 +57,18 @@ class ForwardTafel(BaseKinetics):
         ) = self._get_interface_variables_for_first_order(variables)
         eta_r = delta_phi - ocp
         return (2 * u * j0.diff(c_e)) * pybamm.exp(
-            (ne / (2 * (1 + self.param.Theta * T))) * eta_r
+            (ne * alpha / (2 * (1 + self.param.Theta * T))) * eta_r
         )
 
     def _get_dj_ddeltaphi(self, variables):
         """See :meth:`pybamm.interface.kinetics.BaseKinetics._get_dj_ddeltaphi`"""
+        alpha = self.alpha
         _, delta_phi, j0, ne, ocp, T, u = self._get_interface_variables_for_first_order(
             variables
         )
         eta_r = delta_phi - ocp
         return (2 * u * j0 * (ne / (2 * (1 + self.param.Theta * T)))) * pybamm.exp(
-            (ne / 2) * eta_r
+            (ne * alpha / (2 * (1 + self.param.Theta * T))) * eta_r
         )